Cambio di paramentro per i filtri da sampling_frequency a resolution. Cambiati num_samples in num_elem per ridurre ambiguità.

+++ src/plotwindow.cpp
@@ -15,7 +15,7 @@
 PlotWindow::PlotWindow(QWidget *parent) :
     QMainWindow(parent),
     ui(new Ui::PlotWindow),
-    filter_window(SAMPLE_RATE/NUM_SAMPLES, COMP_SAMPLES, this)
+    filter_window(static_cast<double>(SAMPLE_RATE)/NUM_SAMPLES, COMP_SAMPLES, this)
 {
     ui->setupUi(this);
